Update tmux installer

+++ script/install-tmux.sh
@@ -1,12 +1,13 @@
 #!/usr/bin/env bash
-# Builds tmux and installs for the local user.
+# Builds tmux and installs for the local user, on Debian-based Linux.
 
 set -eu
 
 # Set this to a valid tag name or 'master'.
-TMUX_VERSION="2.7"
+# TMUX_VERSION="2.7"
+TMUX_VERSION="3.1"
 
-sudo apt install --yes --no-install-recommends libevent-2.0.5 libevent-dev
+sudo apt install --yes --no-install-recommends libevent-dev
 
 if [[ -z "$PROJECTS" ]]; then
   echo >&2 "Please run this script after the basic bootstrap setup, so that " \
@@ -21,20 +22,24 @@ if ! [[ -d "$PROJECTS" ]]; then
   exit 20
 fi
 
-if [[ -d "$PROJECTS/tmux" ]]; then
+if [[ -d "$PROJECTS/tmux" && -f "$PROJECTS/tmux/Makefile" ]]; then
   # TODO alternatively, cd, make uninstall, make clean, git pull
   (cd "$PROJECTS/tmux" && make uninstall && make clean)
   mv -v "$PROJECTS/tmux" "$PROJECTS/tmux-old" || exit 25
 fi
 
-cd "$PROJECTS"
-git clone https://github.com/tmux/tmux.git || exit 30
-cd tmux
+if [[ ! -d "$PROJECTS/tmux" ]]; then
+  cd "$PROJECTS"
+  git clone https://github.com/tmux/tmux.git || exit 30
+fi
+
+cd $PROJECTS/tmux
 git fetch && git checkout "$TMUX_VERSION"
 sh autogen.sh || exit 40
 ./configure --prefix "$HOME/.local" || exit 50
-make -j24 || exit 60
+make -j$(nproc) || exit 60
 make install || exit 70
 
+echo
 echo "tmux version [$TMUX_VERSION] installed successfully"
